chore(replicache): extract shared bench utilities into bench-util.ts
Move duplicated benchmark helpers (ReplicachePerfTest, makeRep, populate,
putMap, makeRepWithPopulate, createIndexDefinitions, closeAndCleanupRep,
setupPersistedData, sleep, range, sampleSize, shuffle, valSize) from
replicache.bench.ts and replicache-perf/benchmarks/replicache.ts into a
single packages/replicache/src/bench-util.ts. Also deduplicate the
generateEntries function in btree/write.bench.ts.
